Fixed lag spikes
Turns out that it was actually the RadiusTools class causing lag spikes, generating a random position within a specified area via a while loop uses a lot of cpu time. Instead, for regular uses, such as spawning specials, an array of possible positions is pre-generated when the game starts, one of these vector3s is then chosen at random when required.